Update ShareDesc with realtime post meta value
As the post meta value changes in the editor, we re-render ShareDesc
with the latest desc value.

The function inside useSelect() is automatically fired everytime the
data in WordPress changes (e.g. post meta or even the main body content)
and the select is run to get the latest value.